Bug fixes for Visual Studio 2013 (#144)
This includes several fixes related to windows support.

* Fix compile error related to snprintf on Windows for MSVC < 2015 (1900).

* Add _VARIADIC_MAX=10 definition for MSVC versions older than 2013 (default max for their pseudo-variadic-templates is 5).

* Fix python support for windows.

* Fix python module build structure and unit testing on windows

thanks to Kathleen Biagas who worked though many of these issues!

+++ src/libs/relay/python/conduit_relay_web_python.cpp
@@ -352,7 +352,7 @@ PyRelay_Web_WebServer_set_port(PyRelay_Web_WebServer *self,
         return NULL;
     }
     
-    self->webserver->set_port(port);
+    self->webserver->set_port((int)port);
     
     Py_RETURN_NONE; 
 }
